requirements.txt
flask==2.3.2
tensorflow-cpu==2.12.0
pillow==9.5.0
numpy==1.24.3
gunicorn==20.1.0
gdown==4.7.1